rename QueryModifier namespace
