+++ src/Specification/Specification.php
@@ -3,7 +3,7 @@
 namespace Happyr\DoctrineSpecification\Specification;
 
 use Happyr\DoctrineSpecification\Filter\Filter;
-use Happyr\DoctrineSpecification\Query\QueryModifier;
+use Happyr\DoctrineSpecification\QueryModifier\QueryModifier;
 
 interface Specification extends Filter, QueryModifier
 {
